// backend/services/api-server/src/api/v1/auth/dto/enum.d.ts
export declare enum UserTypes {
    SUPER_ADMIN = -1,
    ADMIN = -2,
    DRIVING_SCHOOL_OWNER = 2,
    DRIVING_SCHOOL_MANAGER = 3
}
export declare enum SystemLogProcessTypes {
    LOGIN = 0,
    LOGOUT = 1,
    FORGOT_PASSWORD = 2,
    RESET_PASSWORD = 3,
    CHANGE_PASSWORD = 4,
    UPDATE_PROFILE = 5,
    UPDATE_EMAIL = 6,
    UPDATE_PHONE = 7
}